Change type to use `ulong` rather than `ushort` for the token index implemented in HtmlData.cs.

+++ source/CsQuery/Dom/IDomObject.cs
@@ -92,7 +92,7 @@ namespace CsQuery
         /// All the unique class names applied to this object.
         /// </summary>
         /// <value>
-        /// A sequence of strings	   
+        /// A sequence of strings      
         /// </value>
 
         IEnumerable<string> Classes { get; }
@@ -736,13 +736,13 @@ namespace CsQuery
         /// have to reindex when removing things)
         /// </summary>
 
-        ushort NodePathID { get; }
+        ulong NodePathID { get; }
 
         /// <summary>
         /// Gets the full pathname of the node file.
         /// </summary>
 
-        ushort[] NodePath { get; }
+        ulong[] NodePath { get; }
 
         /// <summary>
         /// Wrap this element in a CQ object. This is the CsQuery equivalent of the common jQuery
@@ -770,6 +770,6 @@ namespace CsQuery
         /// The internal token ID for this element's node name. 
         /// </summary>
 
-        ushort NodeNameID { get; }
+        ulong NodeNameID { get; }
     }
 }
